refactor(alloc): extract bitmap scratch pool into caliptra-mcu-scratch-alloc
- Move the generic bitmap pool out of the SPDM PAL into a new crate at
  runtime/userspace/scratch-alloc, alongside error/syscall/userlog.
- Non-SPDM tasks (image loader, boot measurement, MCU mailbox) no longer
  depend on caliptra-mcu-spdm-pal just to get scratch memory.
- Only the SpdmPalAlloc impl stays in spdm-pal; the new crate depends
  solely on mcu-error and api-lite (the ApiAlloc/ApiAllocPool impls must
  travel with the type or they violate the orphan rule).
- Pool types are not re-exported from spdm-pal, so the removed dependency
  edge cannot silently reappear.
- Rename McuSpdmBox to ScratchBox; it was never SPDM-specific.
- No behavior or size change: each task still owns its pool and
  BitmapAllocator remains !Send + !Sync.
