+++ platforms/emulator/runtime/userspace/apps/example/Cargo.toml
@@ -25,6 +25,7 @@ caliptra-mcu-libtockasync.workspace = true
 caliptra-mcu-config-emulator.workspace = true
 caliptra-mcu-config-fpga.workspace = true
 caliptra-mcu-pldm-common.workspace = true
+caliptra-mcu-scratch-alloc.workspace = true
 caliptra-mcu-spdm-pal.workspace = true
 mcu-caliptra-api-lite = { workspace = true, features = ["mailbox-io"] }
 portable-atomic.workspace = true
