+++ runtime/userspace/api/spdm-lib/pal/Cargo.toml
@@ -12,6 +12,7 @@ caliptra-mcu-spdm-traits = { workspace = true }
 caliptra-mcu-spdm-codec = { workspace = true }
 caliptra-mcu-measurement-api = { workspace = true }
 mcu-caliptra-api-lite = { workspace = true, features = ["mailbox-io"] }
+caliptra-mcu-scratch-alloc = { workspace = true }
 mcu-error = { workspace = true }
 caliptra-mcu-libsyscall-caliptra = { workspace = true, optional = true }
 caliptra-mcu-libtock_platform = { workspace = true, optional = true }
