+++ runtime/userspace/scratch-alloc/Cargo.toml
@@ -1,0 +1,12 @@
+# Licensed under the Apache-2.0 license
+
+[package]
+name = "caliptra-mcu-scratch-alloc"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+# `ApiAlloc` / `ApiAllocPool` live behind api-lite's `mailbox-io` feature.
+mcu-caliptra-api-lite = { workspace = true, features = ["mailbox-io"] }
+mcu-error = { workspace = true }
